+++ CHANGELOG.md
@@ -15,6 +15,8 @@
 
 - inline/get not-found arity allowed any number of arguments, explicitly changed to one. [#24](https://github.com/bsless/clj-fast/issues/24)
 - Fix callsite analysis of quoted forms. Now functions calls and quoted forms are handled correctly.
+- inline/assoc-in new implementation did not extract bindings. Fixing
+  this allows using side-effecting functions as keys
 
 ### Improve